requirements.txt
flask
flask-cors
pandas
langchain
langchain_experimental
langchain-google-genai
python-dotenv
gunicorn 